API breakage fixed    sync methods based on async methods for iOS and Android added
