+++ source/Xamarin.Auth.LinkSource/AccountStore.cs
@@ -13,8 +13,8 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 //
+using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace Xamarin.Auth
 {
@@ -23,11 +23,11 @@ namespace Xamarin.Auth
 	/// Accounts are stored using a service ID and the username of the account
 	/// as a primary key.
 	/// </summary>
-#if XAMARIN_AUTH_INTERNAL
+	#if XAMARIN_AUTH_INTERNAL
 	internal abstract class AccountStore
-#else
+	#else
 	public abstract partial class AccountStore
-#endif
+	#endif
 	{
 		/// <summary>
 		/// Finds the accounts for a given service.
@@ -38,7 +38,7 @@ namespace Xamarin.Auth
 		/// <param name='serviceId'>
 		/// Service identifier.
 		/// </param>
-		public abstract Task<List<Account>> FindAccountsForServiceAsync (string serviceId);
+		public abstract IEnumerable<Account> FindAccountsForService (string serviceId);
 
 		/// <summary>
 		/// Save the specified account by combining its username and the serviceId
@@ -50,7 +50,7 @@ namespace Xamarin.Auth
 		/// <param name='serviceId'>
 		/// Service identifier.
 		/// </param>
-		public abstract Task SaveAsync (Account account, string serviceId);
+		public abstract void Save (Account account, string serviceId);
 
 		/// <summary>
 		/// Deletes the account for a given serviceId.
@@ -61,7 +61,6 @@ namespace Xamarin.Auth
 		/// <param name='serviceId'>
 		/// Service identifier.
 		/// </param>
-		public abstract Task DeleteAsync (Account account, string serviceId);
+		public abstract void Delete (Account account, string serviceId);
 	}
 }
-
