+++ source/Xamarin.Auth.XamarinIOS-Classic/Xamarin.Auth.XamarinIOS-Classic.csproj
@@ -113,9 +113,6 @@
     <Compile Include="..\Xamarin.Auth.XamarinIOS\PlatformSpecific\FormAuthenticatorController.cs">
       <Link>PlatformSpecific\FormAuthenticatorController.cs</Link>
     </Compile>
-    <Compile Include="..\Xamarin.Auth.XamarinIOS\PlatformSpecific\KeyChainAccountStore.cs">
-      <Link>PlatformSpecific\KeyChainAccountStore.cs</Link>
-    </Compile>
     <Compile Include="..\Xamarin.Auth.XamarinIOS\PlatformSpecific\ProgressLabel.cs">
       <Link>PlatformSpecific\ProgressLabel.cs</Link>
     </Compile>
@@ -130,6 +127,15 @@
     <Compile Include="WebAuthenticator.XamarinIOS.Classic.cs" />
     <Compile Include="..\Xamarin.Auth.XamarinIOS\AccountStore.XamarinIOS.cs">
       <Link>AccountStore.XamarinIOS.cs</Link>
+    </Compile>
+    <Compile Include="..\Xamarin.Auth.XamarinIOS\PlatformSpecific\KeyChainAccountStore.Aync.cs">
+      <Link>PlatformSpecific\KeyChainAccountStore.Aync.cs</Link>
+    </Compile>
+    <Compile Include="..\Xamarin.Auth.LinkSource\AccountStore.Async.cs">
+      <Link>AccountStore.Async.cs</Link>
+    </Compile>
+    <Compile Include="..\Xamarin.Auth.XamarinIOS\PlatformSpecific\KeyChainAccountStore.cs">
+      <Link>PlatformSpecific\KeyChainAccountStore.cs</Link>
     </Compile>
   </ItemGroup>
   <ItemGroup>
